Squash warning.
./base/gsfont.c:1006:12: warning: overflow in implicit constant conversion [-Woverflow]

+++ base/gsfont.c
@@ -1003,7 +1003,7 @@ gs_no_encode_char(gs_font *pfont, gs_char chr, gs_glyph_space_t glyph_space)
 int
 gs_no_decode_glyph(gs_font *pfont, gs_glyph glyph, int ch, ushort *unicode_return, unsigned int length)
 {
-    return GS_NO_CHAR;
+    return (int)GS_NO_CHAR;
 }
 
 /* Dummy glyph enumeration procedure */
